Add playlist selection and activation
Shoe the currently selected playlist in the web UI with an option to
activate any other of the user's available playlists. If a different one
is selected, playlist handling is stopped (if ongoing), the new playlist
is checked to make sure it's a valid one the user can modify, and the
songsMap is cleared. Handling still needs to be activated explicitly.

Also, the .env based playlist is now an optional default playlist. If it
isn't set, starting handling is blocked until a valid playlist was
selected via web UI. In addition, `-no-playlist` parameter skips the
default playlist activation as well.

And finally, made up my mind of /toggle vs /start and /stop endpoints,
and the latter ones have won, no more /toggle now.