Remove CLI parameters and their features
The web UI is now designed to select a playlist and processing starting
point, so a default playlist and autostart makes not much sense anymore,
at least right now.

Let's see when I'll regret the decision and bring the feature back, but
in that case I'd probably prefer to have a `-playlist <id>` parameter.